= Moved `gmObjBuilder` to it's own file
